// src/services/generated/models/TaskStatus.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * 任务状态枚举。
 */
export type TaskStatus = 'pending' | 'running' | 'processing' | 'streaming' | 'succeeded' | 'recoverable_timeout' | 'aspect_ratio_mismatch' | 'failed' | 'cancelled';